+++ front/src/components/DataTable.jsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { DataGrid } from "@mui/x-data-grid";
-import ButtonModal from "./ButtonModal";
+import DeleteModal from "./DeleteModal";
 import Button from "@mui/material/Button";
 
 import { Link as RouterLink } from "react-router-dom";
@@ -20,11 +20,11 @@ export default function DataTable(props) {
       renderCell: (params) => {
         return (
           <>
-            <ButtonModal
+            <DeleteModal
               onClick={props.deleteUser}
               buttonText="Borrar"
               onClickParams={{ id: params.row.id }}
-            ></ButtonModal>
+            ></DeleteModal>
 
             <Button
               data={{
@@ -32,7 +32,7 @@ export default function DataTable(props) {
                 dto: { password: "test2PasswordModifi2cado" },
               }}
               component={RouterLink}
-              to="/userCRUD/actualizar"
+              to={`/userCRUD/${params.row.id}/editar`}
             >
               actualizar
             </Button>
